WebServer/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pytesseract==0.3.8
Pillow==8.4.0
cryptography==3.4.7
requests==2.28.0
opencv-python
python-dotenv
